+++ index.html
@@ -8,13 +8,7 @@
   </head>
   <body>
     <div id="app">
-      <header>
-        <h1>MovieList</h1>
-        <div class="search-box">
-          <input type="text" placeholder="검색" />
-          <button class="search-button">검색</button>
-        </div>
-      </header>
+      <movie-header></movie-header>
       <main>
         <section class="item-view">
           <card-list></card-list>
