+++ src/components/CardList.js
@@ -1,4 +1,4 @@
-class CardList extends HTMLElement {
+export default class CardList extends HTMLElement {
   connectedCallback() {
     this.innerHTML = `
         <h2>지금 인기 있는 영화</h2>
