+++ src/components/MoreButton.js
@@ -1,9 +1,20 @@
-class MoreButton extends HTMLElement {
+export default class MoreButton extends HTMLElement {
   connectedCallback() {
     this.innerHTML = /*html*/ ` 
        <button id="more-button" class="btn primary full-width">
             더 보기
        </button>`;
+
+    this.initEventHandler();
+  }
+
+  initEventHandler() {
+    this.querySelector("#more-button").addEventListener("click", () => {
+      this.dispatchEvent(
+        new CustomEvent("more-button-clicked", { bubbles: true }) //detail :
+      );
+      2;
+    });
   }
 }
