+++ src/index.js
@@ -1,9 +1,10 @@
 import "../templates/reset.css";
 import "../templates/common.css";
 import "../src/App";
-import "./components/MovieCard";
-import "./components/MoreButton";
-import "./components/CardList";
+import MovieHeader from "./components/MovieHeader";
+import MovieCard from "./components/MovieCard";
+import MoreButton from "./components/MoreButton";
+import CardList from "./components/CardList";
 import App from "./App";
 
 new App();
